Add unauthorized page, refactor GameService methods, clean up join games
- Added an unauthorized page which we can redirect to in the case of
unauthorized requests, such as entering a match url that you are not
part of.

- Refactored GameService's method names to be clearer. Methods prefixed
with "send" will simply send an RPC call to the backend, while those
prefixed with "request" will send an RPC call AND return a promise of
the result. The latter is practical when you only want to fetch
something once, such as trying to join a game.

- Refactored JoinGame.tsx to remove some duplicated code.

- Refactored GameBoard to use the interface GameState as given by
  GameService. No need to parse a json string into a js object, you
  immediately get the object from the backend.