Fix player not showing up in lobby, cookie singleton, startup auth
When logging in and going directly to creating a match would result in
an empty lobby and game. This bug was caused by a failure to establish
a connection with the gamehub (due to the user not being logged in).

In order to fix this issue we need to recreate the connection with
the hub using the newly received access token. This turned out to
be easier said than done. In the end the solution involves adding a
change listener to the cookies, and when we get a new access token
we'll recreate the GameService instance (and thus the connection).

In order to persist the cookie change listener I created
CookieService.ts. This was because otherwise when we created a new
Cookies object, it would be another instance from the one App.tsx
subscribes to.

In addition, I've added more rigorous authentication measures in
App.tsx. It will now redirect the user to an error page if the servers
are offline, and correctly wait until connection has been established
otherwise.

MatchPage.tsx also checks if the game service is actually connected
before rendering. If that is not the case, that means the authentication
processed has failed, and it will instead render a redirection page for
the user to log in. This situation appears when a user enters a
match-link directly into the browser, without being authenticated.